+++ java_src/foam/dao/LimitedDAO.java
@@ -16,11 +16,11 @@ public class LimitedDAO
     return this;
   }
 
-  public Sink select_(X x, Sink s, Integer skip, Integer limit, Comparator order, Predicate predicate) {
-    return super.select_(x, s, skip, limit_, order, predicate);
+  public Sink select_(X x, Sink s, Long skip, Long limit, Comparator order, Predicate predicate) {
+    return super.select_(x, s, skip, Long.valueOf(limit_), order, predicate);
   }
 
-  public void removeAll_(X x, Integer skip, Integer limit, Comparator order, Predicate predicate) {
-    super.removeAll_(x, skip, limit_, order, predicate);
+  public void removeAll_(X x, Long skip, Long limit, Comparator order, Predicate predicate) {
+    super.removeAll_(x, skip, Long.valueOf(limit_), order, predicate);
   }
 }
